History is completed
History aspect is completed, along with correct slide transitions
between views.  Going to the previous dish when there are no items in
the history also correctly slides you back to the main page.  Could
possibly make an alt text for that previous button to have it say like
"Go back to menu".